requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.22
pypdf>=3.17.0
sentence-transformers>=2.2.2
openai>=1.6.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
